+++ src/common/js/configure.js
@@ -2,22 +2,29 @@
  * pwix:modal/src/common/js/configure.js
  */
 
+import _ from 'lodash';
+
 pwixModal._defaults = {
     verbosity: MD_VERBOSE_NONE
 };
 
-pwixModal._conf = { ...pwixModal._defaults };
+_.merge( pwixModal._conf, pwixModal._defaults );
 
 /**
  * @summary Package configuration
+ *  Should be called *in same terms* both by the client and the server.
  * @locus Anywhere
  * @param {Object} o the runtime configuration of the package
- *  Should be *in same terms* called both by the client and the server.
+ * @returns {Object} the package configuration
  */
 pwixModal.configure = function( o ){
-    pwixModal._conf = { ...pwixModal._defaults, ...o };
-
-    if( pwixModal._conf.verbosity & MD_VERBOSE_CONFIGURE ){
-        console.log( 'pwix:modal configure() with', o, 'building', pwixModal._conf );
+    if( o && _.isObject( o )){
+        _.merge( pwixModal._conf, pwixModal._defaults, o );
+        // be verbose if asked for
+        if( pwixModal._conf.verbosity & MD_VERBOSE_CONFIGURE ){
+            console.log( 'pwix:modal configure() with', o, 'building', pwixModal._conf );
+        }
     }
+    // also acts as a getter
+    return pwixModal._conf;
 };