+++ src/common/js/global.js
@@ -3,4 +3,5 @@
  */
 
 pwixModal = {
+    _conf: {}
 };